pyproject.toml
[project]
name = "platonic-init"
version = "0.1.0"
description = "Experiments for low-dimensional platonic initialization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.3.0",
  "transformers>=4.46.0",
  "trl>=0.11.0",
  "datasets>=2.20.0",
  "accelerate>=0.34.0",
  "safetensors>=0.4.5",
  "numpy>=1.26.0",
  "scipy>=1.12.0",
  "scikit-learn>=1.5.0",
  "tqdm>=4.66.0",
  "pyyaml>=6.0.2",
]

[project.optional-dependencies]
analysis = [
  "matplotlib>=3.9.0",
  "seaborn>=0.13.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true
